routes and logger setup
